api/requirements.txt
flask
marshmallow
Flask-SQLAlchemy
PyJWT
passlib
argon2-cffi
Flask-Migrate
psycopg2
Flask-CORS
gunicorn